+++ src/cleaners/cleaner.test.js
@@ -1,4 +1,5 @@
 import {getAPI, parseFunc, getOpeningScroll, getPeople, getHomeWorld, getSpecies} from './cleaner.js'
+import {mockDirtyFilmData} from './cleaner.js'
 
 describe('cleaner', () => {
 
@@ -21,22 +22,6 @@ describe('cleaner', () => {
 
     it('should call fetch on the url passed to it', () => {
       const url = 'https://swapi.co/api/films/'
-      const mockDirtyFilmData = {
-        characters : ["https://swapi.co/api/people/1/", "https://swapi.co/api/people/2/", "https://swapi.co/api/people/3/", "https://swapi.co/api/people/4/", "https://swapi.co/api/people/5/", "https://swapi.co/api/people/6/", "https://swapi.co/api/people/7/", "https://swapi.co/api/people/8/", "https://swapi.co/api/people/9/", "https://swapi.co/api/people/10/", "https://swapi.co/api/people/12/", "https://swapi.co/api/people/13/", "https://swapi.co/api/people/14/", "https://swapi.co/api/people/15/", "https://swapi.co/api/people/16/", "https://swapi.co/api/people/18/", "https://swapi.co/api/people/19/", "https://swapi.co/api/people/81/"],
-        created: "2014-12-10T14:23:31.880000Z",
-        director: "George Lucas",
-        edited: "2015-04-11T09:46:52.774897Z",
-        episode_id: 4,
-        opening_crawl: "It is a period of civil war. Rebel spaceships, striking from a hidden base, have won their first victory against the evil Galactic Empire. During the battle, Rebel spies managed to steal secret plans to the Empire's ultimate weapon, the DEATH STAR, an armored space station with enough power to destroy an entire planet. Pursued by the Empire's sinister agents, Princess Leia races home aboard her starship, custodian of the stolen plans that can save her people and restore freedom to the galaxy....",
-        planets: ["https://swapi.co/api/planets/2/", "https://swapi.co/api/planets/3/", "https://swapi.co/api/planets/1/"],
-        producer: "Gary Kurtz, Rick McCallum",
-        release_date: "1977-05-25",
-        species: ["https://swapi.co/api/species/5/", "https://swapi.co/api/species/3/", "https://swapi.co/api/species/2/", "https://swapi.co/api/species/1/", "https://swapi.co/api/species/4/"],
-        starships: ["https://swapi.co/api/starships/2/", "https://swapi.co/api/starships/3/", "https://swapi.co/api/starships/5/", "https://swapi.co/api/starships/9/", "https://swapi.co/api/starships/10/", "https://swapi.co/api/starships/11/", "https://swapi.co/api/starships/12/", "https://swapi.co/api/starships/13/"],
-        title: "A New Hope",
-        url: "https://swapi.co/api/films/1/",
-        vehicles: ["https://swapi.co/api/vehicles/4/", "https://swapi.co/api/vehicles/6/", "https://swapi.co/api/vehicles/7/", "https://swapi.co/api/vehicles/8/"],
-      }
       
       window.fetch = jest.fn().mockImplementation(() => ({
         status: 200,
